as.fasta

functionality to convert objects to a fasta string. Currently this can handle character objects which are interpreted as filenames or several of the popular means of storing sequence data: DNAbin, DNAStringSet, AAStringSet, RNAStringSet, BStringSet, DNAMultipleAlignment, RNAMultipleAlignment, or AAMultipleAlignment.

as.fasta(seqs)

Arguments

seqs

(Required.) the sequence/alignment to be displayed. A character vector, DNAbin, DNAStringSet, AAStringSet, or RNAStringSet.

Value

A character string in fasta format.

Examples

seqfile <- system.file("sequences","AHBA.aln",package="msaR")
as.fasta(seqfile)

#> [1] ">Hygrocin_AHBA\n--atg--------------------------------------------------------agcaatgatgtgcgactgcga-------tccgagctgcccgcatggccgcagtacggcgacgaggagcgcgaggccctcatccgggctctggatcaggggcaatggtggcgtatcgggggcggtgaggtcgacgccttcgaggcggagttcgccgcggcccatgggagcgagcacgccctggcggtcaccaatgggacgcatgcgctggagctcgccctcgaagtgctcggggtcggcaccgactccgaggtgatcgttcccgcgttcaccttcatctcgtcctcgcaggcggctcagcggctgggcgcggtggccgtgcccgtggacgtggacccggacacgtactgcatcgatccctcggcggtcgaggcagccatcggcccgaaaacccgcgcgatcatgccggtgcacatggcgggccagatgtgcgacatggacgcgctgggcaagctgtccgccgactcgggggtgccgctgatccaggacgcggcccatgcgcacggtgcgcggtggcgcggtcagaaggtcggtgagctgggctcggtcgccgcgttcagcttccagaacggaaagctgatgacggccggtgagggcggcgccgtgctcttccccgatgccgagatgtacgagaggggcttcgtccggcacagctgtggacgtccgcgcaccgaccgcggctacttccaccgcacctcgggctccaacttccggctgaacgagttctccgcatccgtactgcgcgcccaactcgcccgcctggacggccagatcaccacgcgtgagcagcgctggccggtgctgagccggctgctcgccgaaatccccggtgtggtaccgcagtcgcgcgacgaccgcggtgaccgcaatccgcactacatggcgatgttccgggtgccgggcatcaccgaggagcgccgcgcgaaggtcgtcgacaccctcatcgagcgcggggtgcccgcgtttgtcgcgttccgcgcggtctaccgtacggacgccttctgggaggtcgcggcgccggatctgacggtggacgaactcgcccgccgctgcccgcactccgaggcgctcacccgcgactgcctttggctgcaccaccgggtgctgctgggcagcgaggagcagatgcacgaagtggccgccgtcgtcgccgatgtgctcgcgggcgcatga\n>Macbecin_AHBA\n--atg-----------------------------------------------------------agcaattctccggacgca----ggcatcgagttcccggcctggccgcagcacgacgacgccgagcgcacggccctgctgcgcgcgctggaccagggccagtggtggcgggtcggcggcagcgaggtcgacgagttcgagcgcgagttcgccgagtaccacggcgcgggccacgcgctcgccgtcaccaacgggacgcacgcgctggagctggcgctccaggtgctcgacgtgggccccggcgccgaggtgatcgtgcccgccttcaccttcatctcgtcctcgcaggccgtgcagcgcctcggcgcggtcgccgtgcccgtggacgtcgaccccgacacctactgcctggacgtcgccgccgccgaggccgccgtcacctcccgcacctcggcgatcatgcccgtgcacatggcgggccagttcgccgacatggaccggctggacaagctgtccacgagcaccggcgtgccggtcctgcaggacgccgcccacgcccacggcgcccactggcgcggcaagcgggtcggcgagctcgggtcgatcgccacgttcagcttccagaacggcaagctgatgacggcgggcgagggcggggccgtgctgttcgaggacgaggcccagtgggagaaggcgttcgtcctgcacagctgcggcaggcccaagggcgaccgcggctacttccacctcacctcgggctcgaacttccggatgaacgagttctccgccgccgtgctgcgcgcccagctcggcaggctggactcgcagatcgcgaccaggcaggcccgctggccggtgctgtccgcgctgctggcggggatcgacggcgtcgtgccgcagaccgtcgacccgcgcagcgaccgcaacccgagctacatggcgatgttccggatgcccggcgtcaccgaggagcggcgcaacgcggtggtggacgagctggtgcggcgcgggatcccggcgttcatggcgttccgggcggtctaccgcacccaggcgttctgggagaccggcgcgcccgacctgaccaccgaggagctggccgcccggtgcccggtgtcggaggagatcacccgcgactgcgtgtggctgcaccaccgggtgctgctgggcaccgaggagcagctgcaccggcttgccgccgtcgtggccgacgtggtcgcgggcgcgtga\n>Ansatrienin_AHBA\n--gtg--------------------------------------------------------cacctgaaggccccgcagaag-------ctggagttccccgcctggccgcagttcgactcgaccgagcgcgaggccctgctgcgcacgctcgacagcggccggtggtggcggatcaccggggacgaggtcgtctcgttcgagcgggagttcgcgcaggcccacggcgcccggcacgcgctcgccgtcaccacgggcacgcacgccctcgaactcgtcctggaggtcatgggcgccgggcctggcaccgaggtgatcgtgcccgcgttcaccttcatctcctcctcgcaggccgcccagcggctcggcgcggtcgtcgtccccgtcgacgtcgacccggacacctactgcgtcgacccggccgccgtgcgggccgcgatcaccgagcgcaccagggcgatcatgcccgtgcacatggcgggccagctcgccgacatggacgcgctgcgcgccctgtccgccgagaccggcgtcccgctgctccaggacgccgcgcacgcccacggcgcgcggtggcgcggcgagcccctcggcgcgctcggctcgatcgccgccttcagcttccagaacggcaagctgatgaccgcgggcgagggcggcgccgtgctgctcccggacggggaccagcacgaggaggcgttcctgcgccacgcctgcgggcgcccgctcgccgaccgggcctaccaccaccgcacggccgggtcgaacttccgcatgaacgagttctccgccgcgctgctgcgcgcccagctcggcaggctccagcagcagaccgacgtgcgcgagcggcactggcccgtgctgtcgcggctgctcgccgaggtgcccggcgtcgtcccgcaggggaccgacccgcgctgcgaccgcaacccgcactacatggccatgttccaggtccccggcatcggcgaggagcgccgcgcccgcctggtcgacgcgctggtcgagcgcggagtcccggcgttcgccgggttccgcgcgatctaccgcaccgccgcgttctgggagcgcgccgcccccgacaccacggtggacgccctcgcccgcgcctgcccggtcacggaagcgctcagcagcgactgcgtgtggctgcaccaccgggtgctgctgggggcggaggagcagatgccgctggtggcggaggccgtcgcctgcgcgctggaggaggcctga\n>Ansamitocin_AHBA\n--gtg--------------------------------------------------------cacctgaaggccccgcagaag-------ctggagttccccgcctggccgcagttcgactcgaccgagcgcgaggccctgctgcgcacgctcgacagcggccggtggtggcggatcaccggggacgaggtcgtctcgttcgagcgggagttcgcgcaggcccacggcgcccggcacgcgctcgccgtcaccacgggcacgcacgccctcgaactcgtcctggaggtcatgggcgccgggcctggcaccgaggtgatcgtgcccgcgttcaccttcatctcctcctcgcaggccgcccagcggctcggcgcggtcgtcgtccccgtcgacgtcgacccggacacctactgcgtcgacccggccgccgtgcgggccgcgatcaccgagcgcaccagggcgatcatgcccgtgcacatggcgggccagctcgccgacatggacgcgctgcgcgccctgtccgccgagaccggcgtcccgctgctccaggacgccgcgcacgcccacggcgcgcggtggcgcggcgagcccctcggcgcgctcggctcgatcgccgccttcagcttccagaacggcaagctgatgaccgcgggcgagggcggcgccgtgctgctcccggacggggaccagcacgaggaggcgttcctgcgccacgcctgcgggcgcccgctcgccgaccgggcctaccaccaccgcacggccgggtcgaacttccgcatgaacgagttctccgccgcgctgctgcgcgcccagctcggcaggctccagcagcagaccgacgtgcgcgagcggcactggcccgtgctgtcgcggctgctcgccgaggtgcccggcgtcgtcccgcaggggaccgacccgcgctgcgaccgcaacccgcactacatggccatgttccaggtccccggcatcggcgaggagcgccgcgcccgcctggtcgacgcgctggtcgagcgcggagtcccggcgttcgccgggttccgcgcgatctaccgcaccgccgcgttctgggagcgcgccgcccccgacaccacggtggacgccctcgcccgcgcctgcccggtcacggaagcgctcagcagcgactgcgtgtggctgcaccaccgggtgctgctgggggcggaggagcagatgccgctggtggcggaggccgtcgcctgcgcgctggaggaggcctga\n>Divergolide_AHBA\n--atgacgagaggttttcaccctggttatcgtgcgggccgagaatcattcgaggagggctggcgaacaatgctggacaaggaactcgacctggatttcccggcgtggccgcagtacgacgacaccgagcgtgccgggctcgtccgtgccctggagcagggacagtggtggcgtatcggcggcagtgaggtggacgagttcgagcgcgagttcgccgaggcgaacggtgccccgcacgcgctcgcggtgacgaccggcacgcacgccctggaactggcactgcaggtgctcggcgtcggccccggcaccgaggtcgtcgtgccggcgttcaccttcatctcgtcgtcgcaggccgcgcagcgactgggcgcggtggccgtcccggtcgacgtcgacctggacacgtactgcctggaccccgaggcggtggcggcggcgatcacgccgcgcaccgcggcgatcatgccggtgcacatggccggtcagatctgcgacatggacgccctcggcaagctgtcggcggactccggcgtcccgctgctgcacgacgccgcgcacgcgcacggcgggcggtggcgcgaccagggcgtgtcggcgctgggcacgatggcggcgttcagcttccagaacggcaagctgatgaccgcgggcgagggcggcgcggtcaccttcccggacagcgagcagtacgagacggcgttcctccggcacagctgcggccgcccccgcacggaccgcacctaccggcaccagacctccggctccaacttccggatgaacgagttcaccgcgagcgtgctgcgcgcccaactggcccggctcgacggccagatcgacacccgggagcagcgctggccggtgctggccggacaactcgcgcggatcacgggcgtgctgccgcaggccaccgacgaccggtgcacccgcaacccgcactacatggccatgttccgggtgcccggcatctccgaacagcagcgcaacgaactggtcgacgcgctggtggcccgcggcctgccggccttcgcggcgttccgcgcgatccatcgcaccaaggcgttctgggagaccggggcgcccgacgaaccggtcgaggcgatcgcccggcggtgcccgaacagcgaggcgctgtccaccgactgcgtctggctgcaccaccggaccctgctcggcaccgaggagcagatgcacgccgtcgccgaggtcgtctccgacgcgctggcggcgtcgtga\n>Rubradirin_AHBA\n--atg--------------------------------------------------------------agtccgcgaccgagg-------ccgacctttccggattggccgcaattcgatgacaccgaacgccgagctcttgatcgtgcgctgtcccagggccagtggtggcgcatgggtggcagcgaggtggacagcttcgagcgggagttcgccggataccacggtgcccaccacgccctcgccgtcaccaacggcacgcacgcgctggagctcgccctgcaggtgctgggtgcgggaccgggcaccgaagtgatcgtgccggcgttcaccttcatctcctcgtcccaggccgctcagcggatcggggccgtcgccgtgccggtggacgtcgatccggagacctacaacatcgacgcgaccgccaccgccgaggcgatcacaccgcggacccgggtgatcatgccggtgcacatggcgggtctcatggccgacatggacgcactggacaagctcgcctccgatgccggcgtccggatccttcaggacgcggcccacgcgcacggtgcccggtggcggggcaagcgcgtgggcgagctgggctcgatcgcggcgttcagcttccagaacggcaagctgatgaccgcgggcgagggcggcgcggtgctgttcttggacccggacgactacgagaaggcattcctgcatcactcctgcggccggccgcgcaccgaccgcaactatcaccaccaggtggccggcaccaacatgcggatgaacgagttctcggccgccgtcctgcgcgctcagctgggccgcctcgacggccagatcgagctgcgcgagcagcggtggcggttgctgtcgcagctcttggcccagatccccggcgtccggccgcagggcggtgatgcgcgggccgaccggaatccgcactacatggcgatgttccggcttcccggctggtcggaggagcgtcgcaacctcctcgtcgatcgcctggtcgacgcggggatcccggcgttcgcgggcttccgcgccatctaccgcaccgcggcgttctgggagaccggcgcacccgaggagtccgtcgacgcggtcgccaagcggtgccccaacgccgacgcgatcagccaggactgcgtctggctgcaccaccgcaccctgctcgcgtcggagcaggctctcaccgacaccgccgccatcgtggccgaccaggtggccgccggatga\n>Napthomycin_AHBA\n--atg--------------------------------------------------------------aacgcgcgaccggca-------ccggaattcccgaagtggccgcagtacgacgacgaagagcggaccggactgatccgggccctggaacaaggccagtggtggcgcatgggcggggaggaagtcgactccttcgagcgtgagttcgccgacttccacggcgccccgcacgccctggccgtcaccaacggcacccaggccctcgaactcgccctccaggtgctgggcgtcggtccggggaccgaggtcatcgtgccggccttcaccttcatctcctcctcccaggccgcccagcgcatcggcgcggtcgccgtccccgtcgacgtcgacccgggcacgtacaacatcgacgtcgccgcggccgccgccgccgtcaccccgcgcacccgggcgatcatgcccgtgcacatggccggtctgccggccgacatggacgccctcggcaagctcgccgccgacaccggcgtatcgatcgtgcaggacgcggcccacgcccagggcgcccgctggcagggcaagcgggtcggcgaactcggttcggtcgccgtgttctccttccagaacggcaagttgatgaccgccggtgagggcggcgccgtcctgttccccgacagcggcctgcacgaggcggcgttcctgcggcacagctgcggtcggccgcgtaccgaccgccactaccggcacgaggtcgcgggcaccaacatgcggctcaacgagttctccgccgccgtgctgcgcgcccagctgcgccgcctgcccgcccagaccgaactgcgcgacaagcggtgggcgctgctgtcccggctgctcggcgccatcgacggcgtcgtaccgcagagcggtgacgcccgcgccgaccagaacacgcactacatggcgatgttccgggtccccggcatctccgaggagacccgtaacgcgctggtcgaccggctggtcgacgccgggctgcccgccttcgccgcgttccgcgccatccaccgcaccgacgccttctgggaggtcggcgcccccgacgagaccgtggagcagatcgccgagcgctgcccgaacaccgaggccatcagcgacgactgcgtctggctgcaccaccgggtgctcctcgccgaggagcgggacctggagatcacggccgagatcatcgcggacgcagtggccgccatatga\n>Neoansalactam_AHBA\nacatg--------------------------------------------------------------aacgcgcgacggacg-------ccagagttccccacctggccgcagtacgacgacggcgagcgcaccggcctgatccgggccctggagcagggccagtggtggcgcatgggaggctcggaggtggactccttcgagagtgagttcgcggacttccacggtgccccacacgctttggccgtcaccaacggcacccacgccctggagttggcgttgcagtgtctgggcgtcgggccgggcaccgaggtcatcgtgccggccttcaccttcatctcctcctcccaggccgctcagcggctgggagcggttgccgtccccgtcgacgtcgatctcgatacctacaacatcgacgtggctgccgcggcttccgccgtcacccccctcaccaaggcgatcatgcctgtgcacatggcggggctcatcgccgacatggacgcgctcggcgaactctccgccgacaccggtgtgcctcttctccaggacgccgcccacgcacacggtgcccgctggcagggcaaacgggtgggcgagttgggtacggtcgcctcgttcagcttccagaacggcaagctgatgaccgccggcgagggcggtgcgctgctcctgcccgacgaggagacctacgaggccgcgttcctgcggcacagttgtggccggtcacgtaccgaccgccgatacatgcaccagaccgccggcaccaacatgcggctcaacgagttctccgcggccgtgctccgcgcccagctgggccgcctcgacgcccagatcacgctccgcgatcagcgctggacgctgctgtcccggctgctcggtgagatcgacggcgtcgtaccccagggcagcgacccgcgcgccgaccggaactcccactacatggcgatgttccggatccccggcatatccgaggaggcccgcaacgccctcgtcgacacgctcgtcgaggccggcctgcccgccttcgccgccttccgggcgatctaccgcaccgacgcgttctgggagacggccgcgcccgacaccaccgtcgacaagctcgccgaaagctgcccgcacaccgaggcgatcagcaccgactgcatctggctgcaccatcgagtgctgctcgcctcggaggaggccctccacaccacagccgagatcatcgccgacgccgtggccgcacggtga\n>Ansalactam_AHBA\n---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------gcgccgcgcatgctttggccgtcaccaacggcacccatgccctggaactggcgctgcagtgtctgggtgtcggtccgggcacggaggtcatcgtgccggccttcactttcatctcctcgtcccaggccgctcagcggctcggcgcggtcgccgtccccgtcgacgtcgatctcgacacgtacaacatcgacgtggctgccgcggcggccgccgtcacccctcgcaccagggtgatcatgcccgtgcacatggcggggctcatcgctgacatggacgcgctcgacaagctctccgcggacaccggcgtgcccgtgctgcaggacgccgcccacgcgcacggtgcccgatggcagggcaaacgggtgggcgagttgggcacggtcgccaccttcagcttccagaacggcaagctgatgaccgccggcgagggtggtgccctgctcttccccgacgaggagacgtacgaggccgccttcctgcggcacagctgtggccggccgcgcaccgaccgccactacatgcaccagaccgccggcaccaacatgcggctcaacgagttctctgcctccgtgctccgcgcccaactgagccgcctcgacgcccagctcacgctgcgcgatgagcgctggacactgctgtcccggctgctcggcaagatcgacggcgtcgtaccccaggggggcgatgcgcgcgcc------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------\n>Arifamycini\n--atg--------------------------------------------------------------aacgcacgacaggcg-------cccgaattcccccaatggccgcagtacgacgacgcagagcgggacggcctgatccgtgcgctggagcagggccagtggtggcgcatgggcggcagcgaggtggacgcgttcgagcacgagttcgccgcgcaccacggcgccgcgcacgcgctggcggccaccaccgggacgcacgcgctcgaactcgccctgcagtgcatgggcgtcgggccgggcaccgaggtgatcgtgccggccttcaccttcatctcctcctcccaggcggcccagcggctcggcgcggtcgcggtcccggtcgacgtcgacgccggcacctacaacctcgacccggcggcggtctccgccgccctcaccccccgcacccgcgtgatcatgccggtgcacatggccgggctcatggcggacatggacgcgctcggcaagatctccgccgagtccggggtgccgctgctgcaggacgccgcgcacgcccacggcgcgcgctggcaggggcggcgcgtcggcgagctcggcggcgtcgcggcgttcagcttccagaacggcaaactgatgaccgcgggcgagggcggcgcggtgctgttcccggacggcgacatgtacgagacggcgttcctgcgccacagttgcggccgcccccgggacgaccgccgctacttccaccggatcgcgggctccaacctgcggatgaacgagttctcggcggcggtgctgcgggcccagctgcgccgcctcgacgagcagatcgcggtccgggacgagcgctggaagctgctgtccccgctgctcgcgcggatcgacggcgtcgtcccgcaggcggacgacgtgcgcgccgaccggcactcgcactacatggcgatgttccgcgtccccgggatcaccgaggagcggcgcaacgccctcgtcgaccgcctcgtcgccgccggcgtgcccgcgttcgccgcgttccgcgcgatccaccgcaccgaggcgttctgggagatcggggcgcccgcggagacgccggacgcgatcggggagcgctgccccaacaccgaggccatcagccgcgactgcgtctggctgcaccaccgcgtcctgctggcgggggagccggagatccgcgccgccgccgagatcatcgccgacctcgtggcctccgcatga\n>Chaxamycin_AHBA\n--atg--------------------------------------------------------------aacgcgcgacaggca-------ccagaattcccccgctggcctcagtacgacgagaccgagcgggacggcctgatccgtgcgctggaacagggccagtggtggcgcatgggcggcgacgaggtcgactccttcgagcgggagttcgccgagcaccacggcgccccccacgcgctggcggtcaccaacggcacccacgcgctggaactcgccctgcaatgcctgggcgtcggcccgggcaccgaagtcatcgtgccggccttcaccttcatctcctcctcgcaggccgcccagcggctcggcgccgtcgccgtccccgtcgacgtccacccggacacctactgcatcgacgccgccgcggtcgccgccgccgtcaccccccgcacccgcgccatcatgccggtgcacatggccgggctcatcgcggacatggacgccctggcgaagatctccgccgacaccggggtccccctgctccaggacgccgctcacgcccacggtgcccgctggcagggcaaacgcgtcggtgaactgggcagcgtcgccgcgttcagcttccagaacggcaagctgatgaccgccggcgagggcggcgccctcctcttcccggacacggagctgtacgaggcggcgttcctgcggcacagctgcggccgcccccgcgacgaccgccgctacctgcaccgcaccgcgggctccaacctgcgcatgaacgagttctccgcggccgtgctgcgcgcccagctgagccgcctcgaccagcagatcgccctgcgggaggagcgctggaccctgctgtcccggctgctcgccgccatcgacggcgtcgtcccgcagggcggcgacgtgcgcgtggaccgcaagtcgcactacatggcgatgttccgggtccccggcctcggcgaggagcggcgcaacgccctggtcgaccggctcgtggaggcgggcctgcccgcgttcgccgcgttccgcgcgatctaccgcaccgacgccttctgggagatcggcgcacccgacgagacccccgacgccatcgccgcccgctgcccccacaccgaagccatcagccaggactgcgtctggctgcaccaccgcgtcctgctcgccggggaagcggagatgcacgccaccgccgagatcatcgccgacctcgtggcccgcgcgtga"

help("as.fasta")

#> Rendering development documentation for 'as.fasta'

if (requireNamespace("Biostrings", quietly = TRUE)) {
   seqs <- Biostrings::readDNAStringSet(seqfile)
   as.fasta(seqs)
}

#> [1] ">Hygrocin_AHBA\n--ATG--------------------------------------------------------AGCAATGATGTGCGACTGCGA-------TCCGAGCTGCCCGCATGGCCGCAGTACGGCGACGAGGAGCGCGAGGCCCTCATCCGGGCTCTGGATCAGGGGCAATGGTGGCGTATCGGGGGCGGTGAGGTCGACGCCTTCGAGGCGGAGTTCGCCGCGGCCCATGGGAGCGAGCACGCCCTGGCGGTCACCAATGGGACGCATGCGCTGGAGCTCGCCCTCGAAGTGCTCGGGGTCGGCACCGACTCCGAGGTGATCGTTCCCGCGTTCACCTTCATCTCGTCCTCGCAGGCGGCTCAGCGGCTGGGCGCGGTGGCCGTGCCCGTGGACGTGGACCCGGACACGTACTGCATCGATCCCTCGGCGGTCGAGGCAGCCATCGGCCCGAAAACCCGCGCGATCATGCCGGTGCACATGGCGGGCCAGATGTGCGACATGGACGCGCTGGGCAAGCTGTCCGCCGACTCGGGGGTGCCGCTGATCCAGGACGCGGCCCATGCGCACGGTGCGCGGTGGCGCGGTCAGAAGGTCGGTGAGCTGGGCTCGGTCGCCGCGTTCAGCTTCCAGAACGGAAAGCTGATGACGGCCGGTGAGGGCGGCGCCGTGCTCTTCCCCGATGCCGAGATGTACGAGAGGGGCTTCGTCCGGCACAGCTGTGGACGTCCGCGCACCGACCGCGGCTACTTCCACCGCACCTCGGGCTCCAACTTCCGGCTGAACGAGTTCTCCGCATCCGTACTGCGCGCCCAACTCGCCCGCCTGGACGGCCAGATCACCACGCGTGAGCAGCGCTGGCCGGTGCTGAGCCGGCTGCTCGCCGAAATCCCCGGTGTGGTACCGCAGTCGCGCGACGACCGCGGTGACCGCAATCCGCACTACATGGCGATGTTCCGGGTGCCGGGCATCACCGAGGAGCGCCGCGCGAAGGTCGTCGACACCCTCATCGAGCGCGGGGTGCCCGCGTTTGTCGCGTTCCGCGCGGTCTACCGTACGGACGCCTTCTGGGAGGTCGCGGCGCCGGATCTGACGGTGGACGAACTCGCCCGCCGCTGCCCGCACTCCGAGGCGCTCACCCGCGACTGCCTTTGGCTGCACCACCGGGTGCTGCTGGGCAGCGAGGAGCAGATGCACGAAGTGGCCGCCGTCGTCGCCGATGTGCTCGCGGGCGCATGA\n>Macbecin_AHBA\n--ATG-----------------------------------------------------------AGCAATTCTCCGGACGCA----GGCATCGAGTTCCCGGCCTGGCCGCAGCACGACGACGCCGAGCGCACGGCCCTGCTGCGCGCGCTGGACCAGGGCCAGTGGTGGCGGGTCGGCGGCAGCGAGGTCGACGAGTTCGAGCGCGAGTTCGCCGAGTACCACGGCGCGGGCCACGCGCTCGCCGTCACCAACGGGACGCACGCGCTGGAGCTGGCGCTCCAGGTGCTCGACGTGGGCCCCGGCGCCGAGGTGATCGTGCCCGCCTTCACCTTCATCTCGTCCTCGCAGGCCGTGCAGCGCCTCGGCGCGGTCGCCGTGCCCGTGGACGTCGACCCCGACACCTACTGCCTGGACGTCGCCGCCGCCGAGGCCGCCGTCACCTCCCGCACCTCGGCGATCATGCCCGTGCACATGGCGGGCCAGTTCGCCGACATGGACCGGCTGGACAAGCTGTCCACGAGCACCGGCGTGCCGGTCCTGCAGGACGCCGCCCACGCCCACGGCGCCCACTGGCGCGGCAAGCGGGTCGGCGAGCTCGGGTCGATCGCCACGTTCAGCTTCCAGAACGGCAAGCTGATGACGGCGGGCGAGGGCGGGGCCGTGCTGTTCGAGGACGAGGCCCAGTGGGAGAAGGCGTTCGTCCTGCACAGCTGCGGCAGGCCCAAGGGCGACCGCGGCTACTTCCACCTCACCTCGGGCTCGAACTTCCGGATGAACGAGTTCTCCGCCGCCGTGCTGCGCGCCCAGCTCGGCAGGCTGGACTCGCAGATCGCGACCAGGCAGGCCCGCTGGCCGGTGCTGTCCGCGCTGCTGGCGGGGATCGACGGCGTCGTGCCGCAGACCGTCGACCCGCGCAGCGACCGCAACCCGAGCTACATGGCGATGTTCCGGATGCCCGGCGTCACCGAGGAGCGGCGCAACGCGGTGGTGGACGAGCTGGTGCGGCGCGGGATCCCGGCGTTCATGGCGTTCCGGGCGGTCTACCGCACCCAGGCGTTCTGGGAGACCGGCGCGCCCGACCTGACCACCGAGGAGCTGGCCGCCCGGTGCCCGGTGTCGGAGGAGATCACCCGCGACTGCGTGTGGCTGCACCACCGGGTGCTGCTGGGCACCGAGGAGCAGCTGCACCGGCTTGCCGCCGTCGTGGCCGACGTGGTCGCGGGCGCGTGA\n>Ansatrienin_AHBA\n--GTG--------------------------------------------------------CACCTGAAGGCCCCGCAGAAG-------CTGGAGTTCCCCGCCTGGCCGCAGTTCGACTCGACCGAGCGCGAGGCCCTGCTGCGCACGCTCGACAGCGGCCGGTGGTGGCGGATCACCGGGGACGAGGTCGTCTCGTTCGAGCGGGAGTTCGCGCAGGCCCACGGCGCCCGGCACGCGCTCGCCGTCACCACGGGCACGCACGCCCTCGAACTCGTCCTGGAGGTCATGGGCGCCGGGCCTGGCACCGAGGTGATCGTGCCCGCGTTCACCTTCATCTCCTCCTCGCAGGCCGCCCAGCGGCTCGGCGCGGTCGTCGTCCCCGTCGACGTCGACCCGGACACCTACTGCGTCGACCCGGCCGCCGTGCGGGCCGCGATCACCGAGCGCACCAGGGCGATCATGCCCGTGCACATGGCGGGCCAGCTCGCCGACATGGACGCGCTGCGCGCCCTGTCCGCCGAGACCGGCGTCCCGCTGCTCCAGGACGCCGCGCACGCCCACGGCGCGCGGTGGCGCGGCGAGCCCCTCGGCGCGCTCGGCTCGATCGCCGCCTTCAGCTTCCAGAACGGCAAGCTGATGACCGCGGGCGAGGGCGGCGCCGTGCTGCTCCCGGACGGGGACCAGCACGAGGAGGCGTTCCTGCGCCACGCCTGCGGGCGCCCGCTCGCCGACCGGGCCTACCACCACCGCACGGCCGGGTCGAACTTCCGCATGAACGAGTTCTCCGCCGCGCTGCTGCGCGCCCAGCTCGGCAGGCTCCAGCAGCAGACCGACGTGCGCGAGCGGCACTGGCCCGTGCTGTCGCGGCTGCTCGCCGAGGTGCCCGGCGTCGTCCCGCAGGGGACCGACCCGCGCTGCGACCGCAACCCGCACTACATGGCCATGTTCCAGGTCCCCGGCATCGGCGAGGAGCGCCGCGCCCGCCTGGTCGACGCGCTGGTCGAGCGCGGAGTCCCGGCGTTCGCCGGGTTCCGCGCGATCTACCGCACCGCCGCGTTCTGGGAGCGCGCCGCCCCCGACACCACGGTGGACGCCCTCGCCCGCGCCTGCCCGGTCACGGAAGCGCTCAGCAGCGACTGCGTGTGGCTGCACCACCGGGTGCTGCTGGGGGCGGAGGAGCAGATGCCGCTGGTGGCGGAGGCCGTCGCCTGCGCGCTGGAGGAGGCCTGA\n>Ansamitocin_AHBA\n--GTG--------------------------------------------------------CACCTGAAGGCCCCGCAGAAG-------CTGGAGTTCCCCGCCTGGCCGCAGTTCGACTCGACCGAGCGCGAGGCCCTGCTGCGCACGCTCGACAGCGGCCGGTGGTGGCGGATCACCGGGGACGAGGTCGTCTCGTTCGAGCGGGAGTTCGCGCAGGCCCACGGCGCCCGGCACGCGCTCGCCGTCACCACGGGCACGCACGCCCTCGAACTCGTCCTGGAGGTCATGGGCGCCGGGCCTGGCACCGAGGTGATCGTGCCCGCGTTCACCTTCATCTCCTCCTCGCAGGCCGCCCAGCGGCTCGGCGCGGTCGTCGTCCCCGTCGACGTCGACCCGGACACCTACTGCGTCGACCCGGCCGCCGTGCGGGCCGCGATCACCGAGCGCACCAGGGCGATCATGCCCGTGCACATGGCGGGCCAGCTCGCCGACATGGACGCGCTGCGCGCCCTGTCCGCCGAGACCGGCGTCCCGCTGCTCCAGGACGCCGCGCACGCCCACGGCGCGCGGTGGCGCGGCGAGCCCCTCGGCGCGCTCGGCTCGATCGCCGCCTTCAGCTTCCAGAACGGCAAGCTGATGACCGCGGGCGAGGGCGGCGCCGTGCTGCTCCCGGACGGGGACCAGCACGAGGAGGCGTTCCTGCGCCACGCCTGCGGGCGCCCGCTCGCCGACCGGGCCTACCACCACCGCACGGCCGGGTCGAACTTCCGCATGAACGAGTTCTCCGCCGCGCTGCTGCGCGCCCAGCTCGGCAGGCTCCAGCAGCAGACCGACGTGCGCGAGCGGCACTGGCCCGTGCTGTCGCGGCTGCTCGCCGAGGTGCCCGGCGTCGTCCCGCAGGGGACCGACCCGCGCTGCGACCGCAACCCGCACTACATGGCCATGTTCCAGGTCCCCGGCATCGGCGAGGAGCGCCGCGCCCGCCTGGTCGACGCGCTGGTCGAGCGCGGAGTCCCGGCGTTCGCCGGGTTCCGCGCGATCTACCGCACCGCCGCGTTCTGGGAGCGCGCCGCCCCCGACACCACGGTGGACGCCCTCGCCCGCGCCTGCCCGGTCACGGAAGCGCTCAGCAGCGACTGCGTGTGGCTGCACCACCGGGTGCTGCTGGGGGCGGAGGAGCAGATGCCGCTGGTGGCGGAGGCCGTCGCCTGCGCGCTGGAGGAGGCCTGA\n>Divergolide_AHBA\n--ATGACGAGAGGTTTTCACCCTGGTTATCGTGCGGGCCGAGAATCATTCGAGGAGGGCTGGCGAACAATGCTGGACAAGGAACTCGACCTGGATTTCCCGGCGTGGCCGCAGTACGACGACACCGAGCGTGCCGGGCTCGTCCGTGCCCTGGAGCAGGGACAGTGGTGGCGTATCGGCGGCAGTGAGGTGGACGAGTTCGAGCGCGAGTTCGCCGAGGCGAACGGTGCCCCGCACGCGCTCGCGGTGACGACCGGCACGCACGCCCTGGAACTGGCACTGCAGGTGCTCGGCGTCGGCCCCGGCACCGAGGTCGTCGTGCCGGCGTTCACCTTCATCTCGTCGTCGCAGGCCGCGCAGCGACTGGGCGCGGTGGCCGTCCCGGTCGACGTCGACCTGGACACGTACTGCCTGGACCCCGAGGCGGTGGCGGCGGCGATCACGCCGCGCACCGCGGCGATCATGCCGGTGCACATGGCCGGTCAGATCTGCGACATGGACGCCCTCGGCAAGCTGTCGGCGGACTCCGGCGTCCCGCTGCTGCACGACGCCGCGCACGCGCACGGCGGGCGGTGGCGCGACCAGGGCGTGTCGGCGCTGGGCACGATGGCGGCGTTCAGCTTCCAGAACGGCAAGCTGATGACCGCGGGCGAGGGCGGCGCGGTCACCTTCCCGGACAGCGAGCAGTACGAGACGGCGTTCCTCCGGCACAGCTGCGGCCGCCCCCGCACGGACCGCACCTACCGGCACCAGACCTCCGGCTCCAACTTCCGGATGAACGAGTTCACCGCGAGCGTGCTGCGCGCCCAACTGGCCCGGCTCGACGGCCAGATCGACACCCGGGAGCAGCGCTGGCCGGTGCTGGCCGGACAACTCGCGCGGATCACGGGCGTGCTGCCGCAGGCCACCGACGACCGGTGCACCCGCAACCCGCACTACATGGCCATGTTCCGGGTGCCCGGCATCTCCGAACAGCAGCGCAACGAACTGGTCGACGCGCTGGTGGCCCGCGGCCTGCCGGCCTTCGCGGCGTTCCGCGCGATCCATCGCACCAAGGCGTTCTGGGAGACCGGGGCGCCCGACGAACCGGTCGAGGCGATCGCCCGGCGGTGCCCGAACAGCGAGGCGCTGTCCACCGACTGCGTCTGGCTGCACCACCGGACCCTGCTCGGCACCGAGGAGCAGATGCACGCCGTCGCCGAGGTCGTCTCCGACGCGCTGGCGGCGTCGTGA\n>Rubradirin_AHBA\n--ATG--------------------------------------------------------------AGTCCGCGACCGAGG-------CCGACCTTTCCGGATTGGCCGCAATTCGATGACACCGAACGCCGAGCTCTTGATCGTGCGCTGTCCCAGGGCCAGTGGTGGCGCATGGGTGGCAGCGAGGTGGACAGCTTCGAGCGGGAGTTCGCCGGATACCACGGTGCCCACCACGCCCTCGCCGTCACCAACGGCACGCACGCGCTGGAGCTCGCCCTGCAGGTGCTGGGTGCGGGACCGGGCACCGAAGTGATCGTGCCGGCGTTCACCTTCATCTCCTCGTCCCAGGCCGCTCAGCGGATCGGGGCCGTCGCCGTGCCGGTGGACGTCGATCCGGAGACCTACAACATCGACGCGACCGCCACCGCCGAGGCGATCACACCGCGGACCCGGGTGATCATGCCGGTGCACATGGCGGGTCTCATGGCCGACATGGACGCACTGGACAAGCTCGCCTCCGATGCCGGCGTCCGGATCCTTCAGGACGCGGCCCACGCGCACGGTGCCCGGTGGCGGGGCAAGCGCGTGGGCGAGCTGGGCTCGATCGCGGCGTTCAGCTTCCAGAACGGCAAGCTGATGACCGCGGGCGAGGGCGGCGCGGTGCTGTTCTTGGACCCGGACGACTACGAGAAGGCATTCCTGCATCACTCCTGCGGCCGGCCGCGCACCGACCGCAACTATCACCACCAGGTGGCCGGCACCAACATGCGGATGAACGAGTTCTCGGCCGCCGTCCTGCGCGCTCAGCTGGGCCGCCTCGACGGCCAGATCGAGCTGCGCGAGCAGCGGTGGCGGTTGCTGTCGCAGCTCTTGGCCCAGATCCCCGGCGTCCGGCCGCAGGGCGGTGATGCGCGGGCCGACCGGAATCCGCACTACATGGCGATGTTCCGGCTTCCCGGCTGGTCGGAGGAGCGTCGCAACCTCCTCGTCGATCGCCTGGTCGACGCGGGGATCCCGGCGTTCGCGGGCTTCCGCGCCATCTACCGCACCGCGGCGTTCTGGGAGACCGGCGCACCCGAGGAGTCCGTCGACGCGGTCGCCAAGCGGTGCCCCAACGCCGACGCGATCAGCCAGGACTGCGTCTGGCTGCACCACCGCACCCTGCTCGCGTCGGAGCAGGCTCTCACCGACACCGCCGCCATCGTGGCCGACCAGGTGGCCGCCGGATGA\n>Napthomycin_AHBA\n--ATG--------------------------------------------------------------AACGCGCGACCGGCA-------CCGGAATTCCCGAAGTGGCCGCAGTACGACGACGAAGAGCGGACCGGACTGATCCGGGCCCTGGAACAAGGCCAGTGGTGGCGCATGGGCGGGGAGGAAGTCGACTCCTTCGAGCGTGAGTTCGCCGACTTCCACGGCGCCCCGCACGCCCTGGCCGTCACCAACGGCACCCAGGCCCTCGAACTCGCCCTCCAGGTGCTGGGCGTCGGTCCGGGGACCGAGGTCATCGTGCCGGCCTTCACCTTCATCTCCTCCTCCCAGGCCGCCCAGCGCATCGGCGCGGTCGCCGTCCCCGTCGACGTCGACCCGGGCACGTACAACATCGACGTCGCCGCGGCCGCCGCCGCCGTCACCCCGCGCACCCGGGCGATCATGCCCGTGCACATGGCCGGTCTGCCGGCCGACATGGACGCCCTCGGCAAGCTCGCCGCCGACACCGGCGTATCGATCGTGCAGGACGCGGCCCACGCCCAGGGCGCCCGCTGGCAGGGCAAGCGGGTCGGCGAACTCGGTTCGGTCGCCGTGTTCTCCTTCCAGAACGGCAAGTTGATGACCGCCGGTGAGGGCGGCGCCGTCCTGTTCCCCGACAGCGGCCTGCACGAGGCGGCGTTCCTGCGGCACAGCTGCGGTCGGCCGCGTACCGACCGCCACTACCGGCACGAGGTCGCGGGCACCAACATGCGGCTCAACGAGTTCTCCGCCGCCGTGCTGCGCGCCCAGCTGCGCCGCCTGCCCGCCCAGACCGAACTGCGCGACAAGCGGTGGGCGCTGCTGTCCCGGCTGCTCGGCGCCATCGACGGCGTCGTACCGCAGAGCGGTGACGCCCGCGCCGACCAGAACACGCACTACATGGCGATGTTCCGGGTCCCCGGCATCTCCGAGGAGACCCGTAACGCGCTGGTCGACCGGCTGGTCGACGCCGGGCTGCCCGCCTTCGCCGCGTTCCGCGCCATCCACCGCACCGACGCCTTCTGGGAGGTCGGCGCCCCCGACGAGACCGTGGAGCAGATCGCCGAGCGCTGCCCGAACACCGAGGCCATCAGCGACGACTGCGTCTGGCTGCACCACCGGGTGCTCCTCGCCGAGGAGCGGGACCTGGAGATCACGGCCGAGATCATCGCGGACGCAGTGGCCGCCATATGA\n>Neoansalactam_AHBA\nACATG--------------------------------------------------------------AACGCGCGACGGACG-------CCAGAGTTCCCCACCTGGCCGCAGTACGACGACGGCGAGCGCACCGGCCTGATCCGGGCCCTGGAGCAGGGCCAGTGGTGGCGCATGGGAGGCTCGGAGGTGGACTCCTTCGAGAGTGAGTTCGCGGACTTCCACGGTGCCCCACACGCTTTGGCCGTCACCAACGGCACCCACGCCCTGGAGTTGGCGTTGCAGTGTCTGGGCGTCGGGCCGGGCACCGAGGTCATCGTGCCGGCCTTCACCTTCATCTCCTCCTCCCAGGCCGCTCAGCGGCTGGGAGCGGTTGCCGTCCCCGTCGACGTCGATCTCGATACCTACAACATCGACGTGGCTGCCGCGGCTTCCGCCGTCACCCCCCTCACCAAGGCGATCATGCCTGTGCACATGGCGGGGCTCATCGCCGACATGGACGCGCTCGGCGAACTCTCCGCCGACACCGGTGTGCCTCTTCTCCAGGACGCCGCCCACGCACACGGTGCCCGCTGGCAGGGCAAACGGGTGGGCGAGTTGGGTACGGTCGCCTCGTTCAGCTTCCAGAACGGCAAGCTGATGACCGCCGGCGAGGGCGGTGCGCTGCTCCTGCCCGACGAGGAGACCTACGAGGCCGCGTTCCTGCGGCACAGTTGTGGCCGGTCACGTACCGACCGCCGATACATGCACCAGACCGCCGGCACCAACATGCGGCTCAACGAGTTCTCCGCGGCCGTGCTCCGCGCCCAGCTGGGCCGCCTCGACGCCCAGATCACGCTCCGCGATCAGCGCTGGACGCTGCTGTCCCGGCTGCTCGGTGAGATCGACGGCGTCGTACCCCAGGGCAGCGACCCGCGCGCCGACCGGAACTCCCACTACATGGCGATGTTCCGGATCCCCGGCATATCCGAGGAGGCCCGCAACGCCCTCGTCGACACGCTCGTCGAGGCCGGCCTGCCCGCCTTCGCCGCCTTCCGGGCGATCTACCGCACCGACGCGTTCTGGGAGACGGCCGCGCCCGACACCACCGTCGACAAGCTCGCCGAAAGCTGCCCGCACACCGAGGCGATCAGCACCGACTGCATCTGGCTGCACCATCGAGTGCTGCTCGCCTCGGAGGAGGCCCTCCACACCACAGCCGAGATCATCGCCGACGCCGTGGCCGCACGGTGA\n>Ansalactam_AHBA\n---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------GCGCCGCGCATGCTTTGGCCGTCACCAACGGCACCCATGCCCTGGAACTGGCGCTGCAGTGTCTGGGTGTCGGTCCGGGCACGGAGGTCATCGTGCCGGCCTTCACTTTCATCTCCTCGTCCCAGGCCGCTCAGCGGCTCGGCGCGGTCGCCGTCCCCGTCGACGTCGATCTCGACACGTACAACATCGACGTGGCTGCCGCGGCGGCCGCCGTCACCCCTCGCACCAGGGTGATCATGCCCGTGCACATGGCGGGGCTCATCGCTGACATGGACGCGCTCGACAAGCTCTCCGCGGACACCGGCGTGCCCGTGCTGCAGGACGCCGCCCACGCGCACGGTGCCCGATGGCAGGGCAAACGGGTGGGCGAGTTGGGCACGGTCGCCACCTTCAGCTTCCAGAACGGCAAGCTGATGACCGCCGGCGAGGGTGGTGCCCTGCTCTTCCCCGACGAGGAGACGTACGAGGCCGCCTTCCTGCGGCACAGCTGTGGCCGGCCGCGCACCGACCGCCACTACATGCACCAGACCGCCGGCACCAACATGCGGCTCAACGAGTTCTCTGCCTCCGTGCTCCGCGCCCAACTGAGCCGCCTCGACGCCCAGCTCACGCTGCGCGATGAGCGCTGGACACTGCTGTCCCGGCTGCTCGGCAAGATCGACGGCGTCGTACCCCAGGGGGGCGATGCGCGCGCC------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------\n>Arifamycini\n--ATG--------------------------------------------------------------AACGCACGACAGGCG-------CCCGAATTCCCCCAATGGCCGCAGTACGACGACGCAGAGCGGGACGGCCTGATCCGTGCGCTGGAGCAGGGCCAGTGGTGGCGCATGGGCGGCAGCGAGGTGGACGCGTTCGAGCACGAGTTCGCCGCGCACCACGGCGCCGCGCACGCGCTGGCGGCCACCACCGGGACGCACGCGCTCGAACTCGCCCTGCAGTGCATGGGCGTCGGGCCGGGCACCGAGGTGATCGTGCCGGCCTTCACCTTCATCTCCTCCTCCCAGGCGGCCCAGCGGCTCGGCGCGGTCGCGGTCCCGGTCGACGTCGACGCCGGCACCTACAACCTCGACCCGGCGGCGGTCTCCGCCGCCCTCACCCCCCGCACCCGCGTGATCATGCCGGTGCACATGGCCGGGCTCATGGCGGACATGGACGCGCTCGGCAAGATCTCCGCCGAGTCCGGGGTGCCGCTGCTGCAGGACGCCGCGCACGCCCACGGCGCGCGCTGGCAGGGGCGGCGCGTCGGCGAGCTCGGCGGCGTCGCGGCGTTCAGCTTCCAGAACGGCAAACTGATGACCGCGGGCGAGGGCGGCGCGGTGCTGTTCCCGGACGGCGACATGTACGAGACGGCGTTCCTGCGCCACAGTTGCGGCCGCCCCCGGGACGACCGCCGCTACTTCCACCGGATCGCGGGCTCCAACCTGCGGATGAACGAGTTCTCGGCGGCGGTGCTGCGGGCCCAGCTGCGCCGCCTCGACGAGCAGATCGCGGTCCGGGACGAGCGCTGGAAGCTGCTGTCCCCGCTGCTCGCGCGGATCGACGGCGTCGTCCCGCAGGCGGACGACGTGCGCGCCGACCGGCACTCGCACTACATGGCGATGTTCCGCGTCCCCGGGATCACCGAGGAGCGGCGCAACGCCCTCGTCGACCGCCTCGTCGCCGCCGGCGTGCCCGCGTTCGCCGCGTTCCGCGCGATCCACCGCACCGAGGCGTTCTGGGAGATCGGGGCGCCCGCGGAGACGCCGGACGCGATCGGGGAGCGCTGCCCCAACACCGAGGCCATCAGCCGCGACTGCGTCTGGCTGCACCACCGCGTCCTGCTGGCGGGGGAGCCGGAGATCCGCGCCGCCGCCGAGATCATCGCCGACCTCGTGGCCTCCGCATGA\n>Chaxamycin_AHBA\n--ATG--------------------------------------------------------------AACGCGCGACAGGCA-------CCAGAATTCCCCCGCTGGCCTCAGTACGACGAGACCGAGCGGGACGGCCTGATCCGTGCGCTGGAACAGGGCCAGTGGTGGCGCATGGGCGGCGACGAGGTCGACTCCTTCGAGCGGGAGTTCGCCGAGCACCACGGCGCCCCCCACGCGCTGGCGGTCACCAACGGCACCCACGCGCTGGAACTCGCCCTGCAATGCCTGGGCGTCGGCCCGGGCACCGAAGTCATCGTGCCGGCCTTCACCTTCATCTCCTCCTCGCAGGCCGCCCAGCGGCTCGGCGCCGTCGCCGTCCCCGTCGACGTCCACCCGGACACCTACTGCATCGACGCCGCCGCGGTCGCCGCCGCCGTCACCCCCCGCACCCGCGCCATCATGCCGGTGCACATGGCCGGGCTCATCGCGGACATGGACGCCCTGGCGAAGATCTCCGCCGACACCGGGGTCCCCCTGCTCCAGGACGCCGCTCACGCCCACGGTGCCCGCTGGCAGGGCAAACGCGTCGGTGAACTGGGCAGCGTCGCCGCGTTCAGCTTCCAGAACGGCAAGCTGATGACCGCCGGCGAGGGCGGCGCCCTCCTCTTCCCGGACACGGAGCTGTACGAGGCGGCGTTCCTGCGGCACAGCTGCGGCCGCCCCCGCGACGACCGCCGCTACCTGCACCGCACCGCGGGCTCCAACCTGCGCATGAACGAGTTCTCCGCGGCCGTGCTGCGCGCCCAGCTGAGCCGCCTCGACCAGCAGATCGCCCTGCGGGAGGAGCGCTGGACCCTGCTGTCCCGGCTGCTCGCCGCCATCGACGGCGTCGTCCCGCAGGGCGGCGACGTGCGCGTGGACCGCAAGTCGCACTACATGGCGATGTTCCGGGTCCCCGGCCTCGGCGAGGAGCGGCGCAACGCCCTGGTCGACCGGCTCGTGGAGGCGGGCCTGCCCGCGTTCGCCGCGTTCCGCGCGATCTACCGCACCGACGCCTTCTGGGAGATCGGCGCACCCGACGAGACCCCCGACGCCATCGCCGCCCGCTGCCCCCACACCGAAGCCATCAGCCAGGACTGCGTCTGGCTGCACCACCGCGTCCTGCTCGCCGGGGAAGCGGAGATGCACGCCACCGCCGAGATCATCGCCGACCTCGTGGCCCGCGCGTGA"