as.sequences

functionality to convert sequence objects into R lists that can be serialized to JS as JSON. Currently, this can handle character objects which are interpreted as filenames or several of the popular means of storing sequence data: DNAbin, DNAStringSet, AAStringSet, RNAStringSet, BStringSet, DNAMultipleAlignment, RNAMultipleAlignment, or AAMultipleAlignment.

as.sequences(seqs)

Arguments

seqs
(Required.) the sequence/alignment to be displayed. A character vector, DNAbin, DNAStringSet, AAStringSet, or RNAStringSet.

Value

A list of named lists where each sublist has name, id, and seq members.

Examples

seqfile <- system.file("sequences","AHBA.aln",package="msaR")
as.sequences(seqfile)

#> [[1]]
#> [[1]]$id
#> [1] "Hygrocin_AHBA"
#> 
#> [[1]]$name
#> [1] "Hygrocin_AHBA"
#> 
#> [[1]]$seq
#> [1] "--atg--------------------------------------------------------agcaatgatgtgcgactgcga-------tccgagctgcccgcatggccgcagtacggcgacgaggagcgcgaggccctcatccgggctctggatcaggggcaatggtggcgtatcgggggcggtgaggtcgacgccttcgaggcggagttcgccgcggcccatgggagcgagcacgccctggcggtcaccaatgggacgcatgcgctggagctcgccctcgaagtgctcggggtcggcaccgactccgaggtgatcgttcccgcgttcaccttcatctcgtcctcgcaggcggctcagcggctgggcgcggtggccgtgcccgtggacgtggacccggacacgtactgcatcgatccctcggcggtcgaggcagccatcggcccgaaaacccgcgcgatcatgccggtgcacatggcgggccagatgtgcgacatggacgcgctgggcaagctgtccgccgactcgggggtgccgctgatccaggacgcggcccatgcgcacggtgcgcggtggcgcggtcagaaggtcggtgagctgggctcggtcgccgcgttcagcttccagaacggaaagctgatgacggccggtgagggcggcgccgtgctcttccccgatgccgagatgtacgagaggggcttcgtccggcacagctgtggacgtccgcgcaccgaccgcggctacttccaccgcacctcgggctccaacttccggctgaacgagttctccgcatccgtactgcgcgcccaactcgcccgcctggacggccagatcaccacgcgtgagcagcgctggccggtgctgagccggctgctcgccgaaatccccggtgtggtaccgcagtcgcgcgacgaccgcggtgaccgcaatccgcactacatggcgatgttccgggtgccgggcatcaccgaggagcgccgcgcgaaggtcgtcgacaccctcatcgagcgcggggtgcccgcgtttgtcgcgttccgcgcggtctaccgtacggacgccttctgggaggtcgcggcgccggatctgacggtggacgaactcgcccgccgctgcccgcactccgaggcgctcacccgcgactgcctttggctgcaccaccgggtgctgctgggcagcgaggagcagatgcacgaagtggccgccgtcgtcgccgatgtgctcgcgggcgcatga"
#> 
#> 
#> [[2]]
#> [[2]]$id
#> [1] "Macbecin_AHBA"
#> 
#> [[2]]$name
#> [1] "Macbecin_AHBA"
#> 
#> [[2]]$seq
#> [1] "--atg-----------------------------------------------------------agcaattctccggacgca----ggcatcgagttcccggcctggccgcagcacgacgacgccgagcgcacggccctgctgcgcgcgctggaccagggccagtggtggcgggtcggcggcagcgaggtcgacgagttcgagcgcgagttcgccgagtaccacggcgcgggccacgcgctcgccgtcaccaacgggacgcacgcgctggagctggcgctccaggtgctcgacgtgggccccggcgccgaggtgatcgtgcccgccttcaccttcatctcgtcctcgcaggccgtgcagcgcctcggcgcggtcgccgtgcccgtggacgtcgaccccgacacctactgcctggacgtcgccgccgccgaggccgccgtcacctcccgcacctcggcgatcatgcccgtgcacatggcgggccagttcgccgacatggaccggctggacaagctgtccacgagcaccggcgtgccggtcctgcaggacgccgcccacgcccacggcgcccactggcgcggcaagcgggtcggcgagctcgggtcgatcgccacgttcagcttccagaacggcaagctgatgacggcgggcgagggcggggccgtgctgttcgaggacgaggcccagtgggagaaggcgttcgtcctgcacagctgcggcaggcccaagggcgaccgcggctacttccacctcacctcgggctcgaacttccggatgaacgagttctccgccgccgtgctgcgcgcccagctcggcaggctggactcgcagatcgcgaccaggcaggcccgctggccggtgctgtccgcgctgctggcggggatcgacggcgtcgtgccgcagaccgtcgacccgcgcagcgaccgcaacccgagctacatggcgatgttccggatgcccggcgtcaccgaggagcggcgcaacgcggtggtggacgagctggtgcggcgcgggatcccggcgttcatggcgttccgggcggtctaccgcacccaggcgttctgggagaccggcgcgcccgacctgaccaccgaggagctggccgcccggtgcccggtgtcggaggagatcacccgcgactgcgtgtggctgcaccaccgggtgctgctgggcaccgaggagcagctgcaccggcttgccgccgtcgtggccgacgtggtcgcgggcgcgtga"
#> 
#> 
#> [[3]]
#> [[3]]$id
#> [1] "Ansatrienin_AHBA"
#> 
#> [[3]]$name
#> [1] "Ansatrienin_AHBA"
#> 
#> [[3]]$seq
#> [1] "--gtg--------------------------------------------------------cacctgaaggccccgcagaag-------ctggagttccccgcctggccgcagttcgactcgaccgagcgcgaggccctgctgcgcacgctcgacagcggccggtggtggcggatcaccggggacgaggtcgtctcgttcgagcgggagttcgcgcaggcccacggcgcccggcacgcgctcgccgtcaccacgggcacgcacgccctcgaactcgtcctggaggtcatgggcgccgggcctggcaccgaggtgatcgtgcccgcgttcaccttcatctcctcctcgcaggccgcccagcggctcggcgcggtcgtcgtccccgtcgacgtcgacccggacacctactgcgtcgacccggccgccgtgcgggccgcgatcaccgagcgcaccagggcgatcatgcccgtgcacatggcgggccagctcgccgacatggacgcgctgcgcgccctgtccgccgagaccggcgtcccgctgctccaggacgccgcgcacgcccacggcgcgcggtggcgcggcgagcccctcggcgcgctcggctcgatcgccgccttcagcttccagaacggcaagctgatgaccgcgggcgagggcggcgccgtgctgctcccggacggggaccagcacgaggaggcgttcctgcgccacgcctgcgggcgcccgctcgccgaccgggcctaccaccaccgcacggccgggtcgaacttccgcatgaacgagttctccgccgcgctgctgcgcgcccagctcggcaggctccagcagcagaccgacgtgcgcgagcggcactggcccgtgctgtcgcggctgctcgccgaggtgcccggcgtcgtcccgcaggggaccgacccgcgctgcgaccgcaacccgcactacatggccatgttccaggtccccggcatcggcgaggagcgccgcgcccgcctggtcgacgcgctggtcgagcgcggagtcccggcgttcgccgggttccgcgcgatctaccgcaccgccgcgttctgggagcgcgccgcccccgacaccacggtggacgccctcgcccgcgcctgcccggtcacggaagcgctcagcagcgactgcgtgtggctgcaccaccgggtgctgctgggggcggaggagcagatgccgctggtggcggaggccgtcgcctgcgcgctggaggaggcctga"
#> 
#> 
#> [[4]]
#> [[4]]$id
#> [1] "Ansamitocin_AHBA"
#> 
#> [[4]]$name
#> [1] "Ansamitocin_AHBA"
#> 
#> [[4]]$seq
#> [1] "--gtg--------------------------------------------------------cacctgaaggccccgcagaag-------ctggagttccccgcctggccgcagttcgactcgaccgagcgcgaggccctgctgcgcacgctcgacagcggccggtggtggcggatcaccggggacgaggtcgtctcgttcgagcgggagttcgcgcaggcccacggcgcccggcacgcgctcgccgtcaccacgggcacgcacgccctcgaactcgtcctggaggtcatgggcgccgggcctggcaccgaggtgatcgtgcccgcgttcaccttcatctcctcctcgcaggccgcccagcggctcggcgcggtcgtcgtccccgtcgacgtcgacccggacacctactgcgtcgacccggccgccgtgcgggccgcgatcaccgagcgcaccagggcgatcatgcccgtgcacatggcgggccagctcgccgacatggacgcgctgcgcgccctgtccgccgagaccggcgtcccgctgctccaggacgccgcgcacgcccacggcgcgcggtggcgcggcgagcccctcggcgcgctcggctcgatcgccgccttcagcttccagaacggcaagctgatgaccgcgggcgagggcggcgccgtgctgctcccggacggggaccagcacgaggaggcgttcctgcgccacgcctgcgggcgcccgctcgccgaccgggcctaccaccaccgcacggccgggtcgaacttccgcatgaacgagttctccgccgcgctgctgcgcgcccagctcggcaggctccagcagcagaccgacgtgcgcgagcggcactggcccgtgctgtcgcggctgctcgccgaggtgcccggcgtcgtcccgcaggggaccgacccgcgctgcgaccgcaacccgcactacatggccatgttccaggtccccggcatcggcgaggagcgccgcgcccgcctggtcgacgcgctggtcgagcgcggagtcccggcgttcgccgggttccgcgcgatctaccgcaccgccgcgttctgggagcgcgccgcccccgacaccacggtggacgccctcgcccgcgcctgcccggtcacggaagcgctcagcagcgactgcgtgtggctgcaccaccgggtgctgctgggggcggaggagcagatgccgctggtggcggaggccgtcgcctgcgcgctggaggaggcctga"
#> 
#> 
#> [[5]]
#> [[5]]$id
#> [1] "Divergolide_AHBA"
#> 
#> [[5]]$name
#> [1] "Divergolide_AHBA"
#> 
#> [[5]]$seq
#> [1] "--atgacgagaggttttcaccctggttatcgtgcgggccgagaatcattcgaggagggctggcgaacaatgctggacaaggaactcgacctggatttcccggcgtggccgcagtacgacgacaccgagcgtgccgggctcgtccgtgccctggagcagggacagtggtggcgtatcggcggcagtgaggtggacgagttcgagcgcgagttcgccgaggcgaacggtgccccgcacgcgctcgcggtgacgaccggcacgcacgccctggaactggcactgcaggtgctcggcgtcggccccggcaccgaggtcgtcgtgccggcgttcaccttcatctcgtcgtcgcaggccgcgcagcgactgggcgcggtggccgtcccggtcgacgtcgacctggacacgtactgcctggaccccgaggcggtggcggcggcgatcacgccgcgcaccgcggcgatcatgccggtgcacatggccggtcagatctgcgacatggacgccctcggcaagctgtcggcggactccggcgtcccgctgctgcacgacgccgcgcacgcgcacggcgggcggtggcgcgaccagggcgtgtcggcgctgggcacgatggcggcgttcagcttccagaacggcaagctgatgaccgcgggcgagggcggcgcggtcaccttcccggacagcgagcagtacgagacggcgttcctccggcacagctgcggccgcccccgcacggaccgcacctaccggcaccagacctccggctccaacttccggatgaacgagttcaccgcgagcgtgctgcgcgcccaactggcccggctcgacggccagatcgacacccgggagcagcgctggccggtgctggccggacaactcgcgcggatcacgggcgtgctgccgcaggccaccgacgaccggtgcacccgcaacccgcactacatggccatgttccgggtgcccggcatctccgaacagcagcgcaacgaactggtcgacgcgctggtggcccgcggcctgccggccttcgcggcgttccgcgcgatccatcgcaccaaggcgttctgggagaccggggcgcccgacgaaccggtcgaggcgatcgcccggcggtgcccgaacagcgaggcgctgtccaccgactgcgtctggctgcaccaccggaccctgctcggcaccgaggagcagatgcacgccgtcgccgaggtcgtctccgacgcgctggcggcgtcgtga"
#> 
#> 
#> [[6]]
#> [[6]]$id
#> [1] "Rubradirin_AHBA"
#> 
#> [[6]]$name
#> [1] "Rubradirin_AHBA"
#> 
#> [[6]]$seq
#> [1] "--atg--------------------------------------------------------------agtccgcgaccgagg-------ccgacctttccggattggccgcaattcgatgacaccgaacgccgagctcttgatcgtgcgctgtcccagggccagtggtggcgcatgggtggcagcgaggtggacagcttcgagcgggagttcgccggataccacggtgcccaccacgccctcgccgtcaccaacggcacgcacgcgctggagctcgccctgcaggtgctgggtgcgggaccgggcaccgaagtgatcgtgccggcgttcaccttcatctcctcgtcccaggccgctcagcggatcggggccgtcgccgtgccggtggacgtcgatccggagacctacaacatcgacgcgaccgccaccgccgaggcgatcacaccgcggacccgggtgatcatgccggtgcacatggcgggtctcatggccgacatggacgcactggacaagctcgcctccgatgccggcgtccggatccttcaggacgcggcccacgcgcacggtgcccggtggcggggcaagcgcgtgggcgagctgggctcgatcgcggcgttcagcttccagaacggcaagctgatgaccgcgggcgagggcggcgcggtgctgttcttggacccggacgactacgagaaggcattcctgcatcactcctgcggccggccgcgcaccgaccgcaactatcaccaccaggtggccggcaccaacatgcggatgaacgagttctcggccgccgtcctgcgcgctcagctgggccgcctcgacggccagatcgagctgcgcgagcagcggtggcggttgctgtcgcagctcttggcccagatccccggcgtccggccgcagggcggtgatgcgcgggccgaccggaatccgcactacatggcgatgttccggcttcccggctggtcggaggagcgtcgcaacctcctcgtcgatcgcctggtcgacgcggggatcccggcgttcgcgggcttccgcgccatctaccgcaccgcggcgttctgggagaccggcgcacccgaggagtccgtcgacgcggtcgccaagcggtgccccaacgccgacgcgatcagccaggactgcgtctggctgcaccaccgcaccctgctcgcgtcggagcaggctctcaccgacaccgccgccatcgtggccgaccaggtggccgccggatga"
#> 
#> 
#> [[7]]
#> [[7]]$id
#> [1] "Napthomycin_AHBA"
#> 
#> [[7]]$name
#> [1] "Napthomycin_AHBA"
#> 
#> [[7]]$seq
#> [1] "--atg--------------------------------------------------------------aacgcgcgaccggca-------ccggaattcccgaagtggccgcagtacgacgacgaagagcggaccggactgatccgggccctggaacaaggccagtggtggcgcatgggcggggaggaagtcgactccttcgagcgtgagttcgccgacttccacggcgccccgcacgccctggccgtcaccaacggcacccaggccctcgaactcgccctccaggtgctgggcgtcggtccggggaccgaggtcatcgtgccggccttcaccttcatctcctcctcccaggccgcccagcgcatcggcgcggtcgccgtccccgtcgacgtcgacccgggcacgtacaacatcgacgtcgccgcggccgccgccgccgtcaccccgcgcacccgggcgatcatgcccgtgcacatggccggtctgccggccgacatggacgccctcggcaagctcgccgccgacaccggcgtatcgatcgtgcaggacgcggcccacgcccagggcgcccgctggcagggcaagcgggtcggcgaactcggttcggtcgccgtgttctccttccagaacggcaagttgatgaccgccggtgagggcggcgccgtcctgttccccgacagcggcctgcacgaggcggcgttcctgcggcacagctgcggtcggccgcgtaccgaccgccactaccggcacgaggtcgcgggcaccaacatgcggctcaacgagttctccgccgccgtgctgcgcgcccagctgcgccgcctgcccgcccagaccgaactgcgcgacaagcggtgggcgctgctgtcccggctgctcggcgccatcgacggcgtcgtaccgcagagcggtgacgcccgcgccgaccagaacacgcactacatggcgatgttccgggtccccggcatctccgaggagacccgtaacgcgctggtcgaccggctggtcgacgccgggctgcccgccttcgccgcgttccgcgccatccaccgcaccgacgccttctgggaggtcggcgcccccgacgagaccgtggagcagatcgccgagcgctgcccgaacaccgaggccatcagcgacgactgcgtctggctgcaccaccgggtgctcctcgccgaggagcgggacctggagatcacggccgagatcatcgcggacgcagtggccgccatatga"
#> 
#> 
#> [[8]]
#> [[8]]$id
#> [1] "Neoansalactam_AHBA"
#> 
#> [[8]]$name
#> [1] "Neoansalactam_AHBA"
#> 
#> [[8]]$seq
#> [1] "acatg--------------------------------------------------------------aacgcgcgacggacg-------ccagagttccccacctggccgcagtacgacgacggcgagcgcaccggcctgatccgggccctggagcagggccagtggtggcgcatgggaggctcggaggtggactccttcgagagtgagttcgcggacttccacggtgccccacacgctttggccgtcaccaacggcacccacgccctggagttggcgttgcagtgtctgggcgtcgggccgggcaccgaggtcatcgtgccggccttcaccttcatctcctcctcccaggccgctcagcggctgggagcggttgccgtccccgtcgacgtcgatctcgatacctacaacatcgacgtggctgccgcggcttccgccgtcacccccctcaccaaggcgatcatgcctgtgcacatggcggggctcatcgccgacatggacgcgctcggcgaactctccgccgacaccggtgtgcctcttctccaggacgccgcccacgcacacggtgcccgctggcagggcaaacgggtgggcgagttgggtacggtcgcctcgttcagcttccagaacggcaagctgatgaccgccggcgagggcggtgcgctgctcctgcccgacgaggagacctacgaggccgcgttcctgcggcacagttgtggccggtcacgtaccgaccgccgatacatgcaccagaccgccggcaccaacatgcggctcaacgagttctccgcggccgtgctccgcgcccagctgggccgcctcgacgcccagatcacgctccgcgatcagcgctggacgctgctgtcccggctgctcggtgagatcgacggcgtcgtaccccagggcagcgacccgcgcgccgaccggaactcccactacatggcgatgttccggatccccggcatatccgaggaggcccgcaacgccctcgtcgacacgctcgtcgaggccggcctgcccgccttcgccgccttccgggcgatctaccgcaccgacgcgttctgggagacggccgcgcccgacaccaccgtcgacaagctcgccgaaagctgcccgcacaccgaggcgatcagcaccgactgcatctggctgcaccatcgagtgctgctcgcctcggaggaggccctccacaccacagccgagatcatcgccgacgccgtggccgcacggtga"
#> 
#> 
#> [[9]]
#> [[9]]$id
#> [1] "Ansalactam_AHBA"
#> 
#> [[9]]$name
#> [1] "Ansalactam_AHBA"
#> 
#> [[9]]$seq
#> [1] "---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------gcgccgcgcatgctttggccgtcaccaacggcacccatgccctggaactggcgctgcagtgtctgggtgtcggtccgggcacggaggtcatcgtgccggccttcactttcatctcctcgtcccaggccgctcagcggctcggcgcggtcgccgtccccgtcgacgtcgatctcgacacgtacaacatcgacgtggctgccgcggcggccgccgtcacccctcgcaccagggtgatcatgcccgtgcacatggcggggctcatcgctgacatggacgcgctcgacaagctctccgcggacaccggcgtgcccgtgctgcaggacgccgcccacgcgcacggtgcccgatggcagggcaaacgggtgggcgagttgggcacggtcgccaccttcagcttccagaacggcaagctgatgaccgccggcgagggtggtgccctgctcttccccgacgaggagacgtacgaggccgccttcctgcggcacagctgtggccggccgcgcaccgaccgccactacatgcaccagaccgccggcaccaacatgcggctcaacgagttctctgcctccgtgctccgcgcccaactgagccgcctcgacgcccagctcacgctgcgcgatgagcgctggacactgctgtcccggctgctcggcaagatcgacggcgtcgtaccccaggggggcgatgcgcgcgcc------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------"
#> 
#> 
#> [[10]]
#> [[10]]$id
#> [1] "Arifamycini"
#> 
#> [[10]]$name
#> [1] "Arifamycini"
#> 
#> [[10]]$seq
#> [1] "--atg--------------------------------------------------------------aacgcacgacaggcg-------cccgaattcccccaatggccgcagtacgacgacgcagagcgggacggcctgatccgtgcgctggagcagggccagtggtggcgcatgggcggcagcgaggtggacgcgttcgagcacgagttcgccgcgcaccacggcgccgcgcacgcgctggcggccaccaccgggacgcacgcgctcgaactcgccctgcagtgcatgggcgtcgggccgggcaccgaggtgatcgtgccggccttcaccttcatctcctcctcccaggcggcccagcggctcggcgcggtcgcggtcccggtcgacgtcgacgccggcacctacaacctcgacccggcggcggtctccgccgccctcaccccccgcacccgcgtgatcatgccggtgcacatggccgggctcatggcggacatggacgcgctcggcaagatctccgccgagtccggggtgccgctgctgcaggacgccgcgcacgcccacggcgcgcgctggcaggggcggcgcgtcggcgagctcggcggcgtcgcggcgttcagcttccagaacggcaaactgatgaccgcgggcgagggcggcgcggtgctgttcccggacggcgacatgtacgagacggcgttcctgcgccacagttgcggccgcccccgggacgaccgccgctacttccaccggatcgcgggctccaacctgcggatgaacgagttctcggcggcggtgctgcgggcccagctgcgccgcctcgacgagcagatcgcggtccgggacgagcgctggaagctgctgtccccgctgctcgcgcggatcgacggcgtcgtcccgcaggcggacgacgtgcgcgccgaccggcactcgcactacatggcgatgttccgcgtccccgggatcaccgaggagcggcgcaacgccctcgtcgaccgcctcgtcgccgccggcgtgcccgcgttcgccgcgttccgcgcgatccaccgcaccgaggcgttctgggagatcggggcgcccgcggagacgccggacgcgatcggggagcgctgccccaacaccgaggccatcagccgcgactgcgtctggctgcaccaccgcgtcctgctggcgggggagccggagatccgcgccgccgccgagatcatcgccgacctcgtggcctccgcatga"
#> 
#> 
#> [[11]]
#> [[11]]$id
#> [1] "Chaxamycin_AHBA"
#> 
#> [[11]]$name
#> [1] "Chaxamycin_AHBA"
#> 
#> [[11]]$seq
#> [1] "--atg--------------------------------------------------------------aacgcgcgacaggca-------ccagaattcccccgctggcctcagtacgacgagaccgagcgggacggcctgatccgtgcgctggaacagggccagtggtggcgcatgggcggcgacgaggtcgactccttcgagcgggagttcgccgagcaccacggcgccccccacgcgctggcggtcaccaacggcacccacgcgctggaactcgccctgcaatgcctgggcgtcggcccgggcaccgaagtcatcgtgccggccttcaccttcatctcctcctcgcaggccgcccagcggctcggcgccgtcgccgtccccgtcgacgtccacccggacacctactgcatcgacgccgccgcggtcgccgccgccgtcaccccccgcacccgcgccatcatgccggtgcacatggccgggctcatcgcggacatggacgccctggcgaagatctccgccgacaccggggtccccctgctccaggacgccgctcacgcccacggtgcccgctggcagggcaaacgcgtcggtgaactgggcagcgtcgccgcgttcagcttccagaacggcaagctgatgaccgccggcgagggcggcgccctcctcttcccggacacggagctgtacgaggcggcgttcctgcggcacagctgcggccgcccccgcgacgaccgccgctacctgcaccgcaccgcgggctccaacctgcgcatgaacgagttctccgcggccgtgctgcgcgcccagctgagccgcctcgaccagcagatcgccctgcgggaggagcgctggaccctgctgtcccggctgctcgccgccatcgacggcgtcgtcccgcagggcggcgacgtgcgcgtggaccgcaagtcgcactacatggcgatgttccgggtccccggcctcggcgaggagcggcgcaacgccctggtcgaccggctcgtggaggcgggcctgcccgcgttcgccgcgttccgcgcgatctaccgcaccgacgccttctgggagatcggcgcacccgacgagacccccgacgccatcgccgcccgctgcccccacaccgaagccatcagccaggactgcgtctggctgcaccaccgcgtcctgctcgccggggaagcggagatgcacgccaccgccgagatcatcgccgacctcgtggcccgcgcgtga"
#> 
#> 

help("as.sequences")

if (FALSE) {
if (requireNamespace("Biostrings")) {
   seqs <- readDNAStringSet(seqfile)
   as.sequences(seqs)
 }
}